+++ sys/arch/alpha/include/cdefs.h
@@ -3,6 +3,6 @@
 #ifndef _MACHINE_CDEFS_H_
 #define	_MACHINE_CDEFS_H_
 
-#define	__ALIGNBYTES		7
+#define	__ALIGNBYTES		((size_t)7)
 
 #endif /* !_MACHINE_CDEFS_H_ */
